Add support for Gradle inference
