+++ pkg/rebuild/maven/mavenstrategy.go
@@ -6,7 +6,6 @@ package maven
 import (
 	"path"
 
-	"github.com/google/oss-rebuild/internal/textwrap"
 	"github.com/pkg/errors"
 
 	"github.com/google/oss-rebuild/pkg/rebuild/flow"
@@ -27,6 +26,7 @@ func (b *MavenBuild) ToWorkflow() (*rebuild.WorkflowStrategy, error) {
 	if !exists {
 		return nil, errors.Errorf("no download URL for JDK version %s", b.JDKVersion)
 	}
+
 	return &rebuild.WorkflowStrategy{
 		Location: b.Location,
 		Source: []flow.Step{{
@@ -43,11 +43,7 @@ func (b *MavenBuild) ToWorkflow() (*rebuild.WorkflowStrategy, error) {
 				Uses: "maven/export-java",
 			},
 			{
-				// TODO: Java 9 needs additional certificate installed in /etc/ssl/certs/java/cacerts
-				// It can be passed to maven command via -Djavax.net.ssl.trustStore=/etc/ssl/certs/java/cacerts
-				Runs: "mvn clean package -DskipTests",
-				// Note `maven` from apt also pull in jdk-21 and hence we must export JAVA_HOME and PATH in the step before
-				Needs: []string{"maven"},
+				Uses: "maven/maven-build",
 			},
 		},
 		OutputDir: path.Join(b.Dir, "target"),
@@ -60,32 +56,4 @@ func (b *MavenBuild) GenerateFor(t rebuild.Target, be rebuild.BuildEnv) (rebuild
 		return rebuild.Instructions{}, err
 	}
 	return workflow.GenerateFor(t, be)
-}
-
-func init() {
-	for _, t := range toolkit {
-		flow.Tools.MustRegister(t)
-	}
-}
-
-var toolkit = []*flow.Tool{
-	{
-		Name: "maven/setup-java",
-		Steps: []flow.Step{
-			{
-				Runs: textwrap.Dedent(`
-				mkdir -p /opt/jdk
-				wget -q -O - "{{.With.versionURL}}" | tar -xzf - --strip-components=1 -C /opt/jdk`[1:]),
-				Needs: []string{"wget"},
-			},
-		},
-	},
-	{
-		Name: "maven/export-java",
-		Steps: []flow.Step{{
-			Runs: textwrap.Dedent(`
-				export JAVA_HOME=/opt/jdk
-				export PATH=$JAVA_HOME/bin:$PATH`[1:]),
-		}},
-	},
 }
